feat: Ex 6.8
Anecdotes, step6

Moved Anecdotes List to its own component.
