+++ part06/redux-anecdotes/src/App.js
@@ -1,31 +1,12 @@
-import { useSelector, useDispatch } from 'react-redux';
 import AnecdotesForm from './components/AnecdotesForm';
-import { voteForAnecdote } from './reducers/anecdoteReducer';
+import AnecdotesList from './components/AnecdotesList';
 
 const App = () => {
-  const anecdotes = useSelector((state) =>
-    state.sort((a, b) => b.votes - a.votes)
-  );
-  const dispatch = useDispatch();
-
-  const vote = (id) => {
-    console.log('vote', id);
-    dispatch(voteForAnecdote(id));
-  };
-
   return (
     <div>
       <h2>Anecdotes</h2>
-      {anecdotes.map((anecdote) => (
-        <div key={anecdote.id}>
-          <div>{anecdote.content}</div>
-          <div>
-            has {anecdote.votes}
-            <button onClick={() => vote(anecdote.id)}>vote</button>
-          </div>
-        </div>
-      ))}
-      <AnecdotesForm></AnecdotesForm>
+      <AnecdotesList />
+      <AnecdotesForm />
     </div>
   );
 };
