+++ part06/redux-anecdotes/src/components/AnecdotesList.js
@@ -1,0 +1,30 @@
+import { useSelector, useDispatch } from 'react-redux';
+import { voteForAnecdote } from '../reducers/anecdoteReducer';
+
+const AnecdotesList = () => {
+  const anecdotes = useSelector((state) =>
+    state.sort((a, b) => b.votes - a.votes)
+  );
+  const dispatch = useDispatch();
+
+  const vote = (id) => {
+    console.log('vote', id);
+    dispatch(voteForAnecdote(id));
+  };
+
+  return (
+    <>
+      {anecdotes.map((anecdote) => (
+        <div key={anecdote.id}>
+          <div>{anecdote.content}</div>
+          <div>
+            has {anecdote.votes}
+            <button onClick={() => vote(anecdote.id)}>vote</button>
+          </div>
+        </div>
+      ))}
+    </>
+  );
+};
+
+export default AnecdotesList;
